+++ matrix-sdk-android/src/main/java/org/matrix/android/sdk/api/session/room/state/StateService.kt
@@ -39,14 +39,11 @@ interface StateService {
     fun updateName(name: String, callback: MatrixCallback<Unit>): Cancelable
 
     /**
-     * Add new alias to the room.
-     */
-    fun addRoomAlias(roomAlias: String, callback: MatrixCallback<Unit>): Cancelable
-
-    /**
      * Update the canonical alias of the room
+     * @param alias the canonical alias, or null to reset the canonical alias of this room
+     * @param altAliases the alternative aliases for this room. It should include the canonical alias if any.
      */
-    fun updateCanonicalAlias(alias: String, callback: MatrixCallback<Unit>): Cancelable
+    fun updateCanonicalAlias(alias: String?, altAliases: List<String>, callback: MatrixCallback<Unit>): Cancelable
 
     /**
      * Update the history readability of the room