+++ vector/src/main/java/im/vector/app/core/di/ScreenComponent.kt
@@ -67,6 +67,7 @@ import im.vector.app.features.roomdirectory.createroom.CreateRoomActivity
 import im.vector.app.features.roommemberprofile.RoomMemberProfileActivity
 import im.vector.app.features.roommemberprofile.devices.DeviceListBottomSheet
 import im.vector.app.features.roomprofile.RoomProfileActivity
+import im.vector.app.features.roomprofile.alias.detail.RoomAliasBottomSheet
 import im.vector.app.features.settings.VectorSettingsActivity
 import im.vector.app.features.settings.devices.DeviceVerificationInfoBottomSheet
 import im.vector.app.features.share.IncomingShareActivity
@@ -153,6 +154,7 @@ interface ScreenComponent {
     fun inject(bottomSheet: ViewEditHistoryBottomSheet)
     fun inject(bottomSheet: DisplayReadReceiptsBottomSheet)
     fun inject(bottomSheet: RoomListQuickActionsBottomSheet)
+    fun inject(bottomSheet: RoomAliasBottomSheet)
     fun inject(bottomSheet: VerificationBottomSheet)
     fun inject(bottomSheet: DeviceVerificationInfoBottomSheet)
     fun inject(bottomSheet: DeviceListBottomSheet)