+++ vector/src/main/java/im/vector/app/features/home/room/list/actions/RoomListQuickActionsBottomSheet.kt
@@ -77,6 +77,7 @@ class RoomListQuickActionsBottomSheet : VectorBaseBottomSheetDialogFragment(), R
 
     override fun onDestroyView() {
         recyclerView.cleanup()
+        roomListActionsEpoxyController.listener = null
         super.onDestroyView()
     }